+++ src/bin/llhd-opt.rs
@@ -1,0 +1,58 @@
+// Copyright (c) 2017-2019 Fabian Schuiki
+
+#[macro_use]
+extern crate clap;
+use clap::Arg;
+use llhd::{assembly::parse_module, pass::ConstantFoldingPass, verifier::Verifier};
+use std::{
+    fs::File,
+    io::{BufWriter, Read},
+    result::Result,
+};
+
+fn main() -> Result<(), String> {
+    let matches = app_from_crate!()
+        .about("Optimizes LLHD assembly.")
+        .arg(
+            Arg::with_name("input")
+                .help("LLHD file to optimize")
+                .required(true),
+        )
+        .arg(
+            Arg::with_name("output")
+                .short("o")
+                .long("output")
+                .takes_value(true)
+                .help("File to write output to; stdout if omitted"),
+        )
+        .get_matches();
+
+    // Read the input.
+    let mut module = {
+        let path = matches.value_of("input").unwrap();
+        let mut input = File::open(path).map_err(|e| format!("{}", e))?;
+        let mut contents = String::new();
+        input
+            .read_to_string(&mut contents)
+            .map_err(|e| format!("{}", e))?;
+        let module = parse_module(&contents).map_err(|e| format!("{}", e))?;
+        let mut verifier = Verifier::new();
+        verifier.verify_module(&module);
+        verifier.finish().map_err(|errs| format!("{}", errs))?;
+        module
+    };
+
+    // Apply optimization pass. Just constant folding for now.
+    ConstantFoldingPass::run_on_module(&mut module);
+
+    // Write the output.
+    if let Some(path) = matches.value_of("output)") {
+        let output = File::create(path).map_err(|e| format!("{}", e))?;
+        let output = BufWriter::with_capacity(1 << 20, output);
+        llhd::assembly::write_module(output, &module);
+    } else {
+        llhd::assembly::write_module(std::io::stdout().lock(), &module);
+    }
+
+    Ok(())
+}